fix(cursor, fetch_dataframe): use column names in cursor's description as is
Previously, the Cursor.fetch_dataframe method lowercased column names preemptively,
thus not respecting the active case-sensitivity configuration value. This could lead
to issues, when e.g. two different columns resolved to the same name when lowercased.

This commit ensures the resulting DataFrame's columns are the same as the ones in the
cursor's description, thus reflecting Redshift's behavior and respecting the underlying
case-sensitivity configuration.